# backend/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: IRIS Enterprise Commerce API
  version: 0.1.0
  description: Source-of-truth contract owned by the backend team.
servers:
  - url: /api
paths:
...
  /products/{id}:
    get:
      operationId: getProductById
      tags: [Catalog]
      parameters:
        - { in: path, name: id, required: true, schema: { type: integer } }
      responses:
        '200':
          description: A single active product.
          content:
            application/json:
              schema:
                type: object
                required: [product]
                properties:
                  product: { $ref: '#/components/schemas/Product' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    cookieAuth: { type: apiKey, in: cookie, name: token }
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
    legacyApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: >-
        DEPRECATED legacy plugin auth (per-user API key + X-Server-Id + IP whitelist). Accepted
        ONLY during the CR-PLUGIN-006 migration overlap window so plugins already in the field do
        not break; each use is logged as a deprecation. New deployments MUST use `hmacAuth`. Will
        be removed once all servers have migrated to signed requests.
    hmacAuth:
      type: apiKey
      in: header
      name: X-Signature
      description: >-
        Signed plugin request authentication (HMAC-SHA256). The signature is carried in
        `X-Signature` and is computed over a canonical string with the per-credential HMAC
        secret. The signing IDENTITY and the signing SECRET are two distinct values:


        Headers (all required):

        - `X-Plugin-Key-Id`: the credential identifier. It IDENTIFIES which signing secret to
          use; it is NOT itself the secret and is safe to log. It is rotatable and revocable
          independently of the secret.

        - `X-Plugin-Version`: plugin build/version string (compatibility + telemetry).

        - `X-Request-Timestamp`: unix epoch milliseconds. Requests outside the clock-skew
          window (+/- 5 minutes / 300000 ms) are rejected.

        - `X-Request-Nonce`: >= 16 chars, single-use; replays within the skew window are
          rejected (nonce cache).

        - `X-Content-Sha256`: hex SHA-256 of the exact JSON request body ('' when no body).

        - `X-Signature`: `HMAC_SHA256(secret, canonicalString)` as lowercase hex, where
          `canonicalString = METHOD \n PATH_AND_QUERY \n TIMESTAMP \n NONCE \n CONTENT_SHA256`.

...
  responses:
...
    NotFound:
      description: Resource not found
  schemas:
...
    Category:
      type: object
      required: [id, name]
      properties:
        id: { type: integer }
        name: { type: string }
        description: { type: [string, 'null'] }
        icon: { type: [string, 'null'] }
        sortOrder: { type: integer }
        isActive: { type: boolean }
        _count:
          type: object
          description: Present on the categories list endpoint.
          properties:
            products: { type: integer }
    Product:
      type: object
      required: [id, name, price, itemBlueprint, quantity, quality, isBlueprint]
      properties:
        id: { type: integer }
        categoryId: { type: [integer, 'null'] }
        name: { type: string }
        description: { type: [string, 'null'] }
        price:
          type: integer
          description: Iris Coin price. Catalog prices are integers; the wallet ledger serializes balances as decimal strings.
        itemBlueprint: { type: string }
        quantity: { type: integer }
        quality: { type: integer, description: '0=Primitive to 5=Ascendant' }
        isBlueprint: { type: boolean }
        imageUrl: { type: [string, 'null'] }
        isActive: { type: boolean }
        isFeatured: { type: boolean }
        sortOrder: { type: integer }
        maxPerUser: { type: [integer, 'null'], description: 'null = unlimited' }
        stock: { type: [integer, 'null'], description: 'null = unlimited' }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        category:
          oneOf:
            - { $ref: '#/components/schemas/Category' }
            - { type: 'null' }
...